+++ core/tasks/timeouts/bulk_timeout_test.go
@@ -38,9 +38,9 @@ func TestBulkTimeout(t *testing.T) {
 	assert.Equal(t, map[string]int{"bulk_timeout": 1}, testsuite.FlushTasks(t, rt, "batch", "throttled"))
 
 	testsuite.AssertContactTasks(t, testdata.Org1, testdata.Cathy, []string{
-		`{"type":"timeout_event","task":{"session_id":123456,"time":"2025-01-01T11:00:00Z","modified_on":"2024-11-15T13:57:00Z"},"queued_on":"2024-11-15T13:59:00Z"}`,
+		`{"type":"timeout_event","task":{"session_id":123456,"modified_on":"2024-11-15T13:57:00Z"},"queued_on":"2024-11-15T13:59:00Z"}`,
 	})
 	testsuite.AssertContactTasks(t, testdata.Org1, testdata.Bob, []string{
-		`{"type":"timeout_event","task":{"session_id":234567,"time":"2025-01-01T12:00:00Z","modified_on":"2024-11-15T13:58:00Z"},"queued_on":"2024-11-15T13:59:00Z"}`,
+		`{"type":"timeout_event","task":{"session_id":234567,"modified_on":"2024-11-15T13:58:00Z"},"queued_on":"2024-11-15T13:59:00Z"}`,
 	})
 }