+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+v10.1.6 (2025-01-21)
+-------------------------
+ * Simplify timeout handling to match expiration handling
+
+v10.1.5 (2025-01-21)
+-------------------------
+ * Handle non-resumable expirations in the contact handler task to ensure correct locking
+
+v10.1.4 (2025-01-20)
+-------------------------
+ * Use session.modified_on to determine if a timed task is still valid
+
 v10.1.3 (2025-01-20)
 -------------------------
  * Assume session.modified_on non-null and include on timed tasks